# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CruxOption {
  id: string;
  label: string;
  probability: number;
  value: string;
  detail: string;
}

export type EvidenceSignal = typeof EvidenceSignal[keyof typeof EvidenceSignal];


export const EvidenceSignal = {
  positive: 'positive',
  neutral: 'neutral',
  negative: 'negative',
} as const;

export interface Evidence {
  source: string;
  detail: string;
  signal: EvidenceSignal;
}

export interface AgentParticipant {
  name: string;
  role: string;
  initials: string;
  color: string;
}

export interface Simulation {
  label: string;
  probability: number;
  revenue: string;
  supportLoad: string;
  accent: string;
}

export type CruxStatus = typeof CruxStatus[keyof typeof CruxStatus];


export const CruxStatus = {
  open: 'open',
  resolved: 'resolved',
  deploying: 'deploying',
  deployed: 'deployed',
} as const;

export interface Crux {
  id: string;
  title: string;
  category: string;
  goal: string;
  owner: string;
  narratedSummary: string;
  urgency: string;
  deadline: string;
  expectedValue: string;
  attention: string;
  cost: string;
  risk: string;
  reversibility: string;
  confidence: number;
  recommendation: string;
  priorityReason: string;
  options: CruxOption[];
  evidence: Evidence[];
  assumptions: string[];
  unresolvedQuestions: string[];
  simulations: Simulation[];
  participants: AgentParticipant[];
  status: CruxStatus;
  /** @nullable */
  selectedResolution?: string | null;
  /** @nullable */
  deploymentStatus?: string | null;
}

export interface ResolutionInput {
  optionId: string;
}

export interface ConversationInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  participant?: string | null;
}

export interface ConversationReply {
  agent: string;
  message: string;
  confidenceShift: number;
}

export interface PortfolioProject {
  name: string;
  code: string;
  invested: string;
  agentHours: number;
  humanHours: number;
  probability: number;
  activeAgents: number;
  unresolved: number;
  progress: number;
  risk: string;
}

export interface Portfolio {
  totalInvested: string;
  totalExpectedValue: string;
  attentionSaved: string;
  projects: PortfolioProject[];
}

export type RunningLoopStatus = typeof RunningLoopStatus[keyof typeof RunningLoopStatus];


export const RunningLoopStatus = {
  running: 'running',
  waiting: 'waiting',
  paused: 'paused',
} as const;

export interface RunningLoop {
  name: string;
  goal: string;
  progress: number;
  spend: string;
  time: string;
  lastAction: string;
  nextMilestone: string;
  status: RunningLoopStatus;
  agents: number;
}

export interface LearnedPolicy {
  name: string;
  description: string;
  decisions: number;
  saved: string;
  enabled: boolean;
}

export interface Learned {
  weeklyDecisions: number;
  weeklySaved: string;
  policies: LearnedPolicy[];
}

export type PreferencesMode = typeof PreferencesMode[keyof typeof PreferencesMode];


export const PreferencesMode = {
  operator: 'operator',
  portfolio: 'portfolio',
  explore: 'explore',
  learn: 'learn',
  contrarian: 'contrarian',
} as const;

export interface Preferences {
  mode: PreferencesMode;
  importance: number;
  urgency: number;
  learning: number;
  novelty: number;
  risk: number;
  cost: number;
}

export type PreferencesInputMode = typeof PreferencesInputMode[keyof typeof PreferencesInputMode];


export const PreferencesInputMode = {
  operator: 'operator',
  portfolio: 'portfolio',
  explore: 'explore',
  learn: 'learn',
  contrarian: 'contrarian',
} as const;

export interface PreferencesInput {
  mode?: PreferencesInputMode;
  importance?: number;
  urgency?: number;
  learning?: number;
  novelty?: number;
  risk?: number;
  cost?: number;
}