# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cruxes/{id}:
    get:
      operationId: getCrux
      tags: [cruxes]
      summary: Get a Crux decision
      parameters:
        - $ref: "#/components/parameters/CruxId"
      responses:
        "200":
          description: Decision detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crux"
        "404":
          description: Crux not found
  /cruxes/{id}/resolve:
    post:
      operationId: resolveCrux
      tags: [cruxes]
      summary: Resolve a Crux and create a deployment artifact
      parameters:
        - $ref: "#/components/parameters/CruxId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolutionInput"
      responses:
        "200":
          description: Resolved decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crux"
  /cruxes/{id}/conversation:
    post:
      operationId: askCrux
      tags: [cruxes]
      summary: Ask an agent about a Crux
      parameters:
        - $ref: "#/components/parameters/CruxId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConversationInput"
      responses:
        "200":
          description: Agent response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationReply"
...
components:
  parameters:
    CruxId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    CruxOption:
      type: object
      required: [id, label, probability, value, detail]
      properties:
        id: { type: string }
        label: { type: string }
        probability: { type: number }
        value: { type: string }
        detail: { type: string }
    Evidence:
      type: object
      required: [source, detail, signal]
      properties:
        source: { type: string }
        detail: { type: string }
        signal: { type: string, enum: [positive, neutral, negative] }
    AgentParticipant:
      type: object
      required: [name, role, initials, color]
      properties:
        name: { type: string }
        role: { type: string }
        initials: { type: string }
        color: { type: string }
    Simulation:
      type: object
      required: [label, probability, revenue, supportLoad, accent]
      properties:
        label: { type: string }
        probability: { type: number }
        revenue: { type: string }
        supportLoad: { type: string }
        accent: { type: string }
    Crux:
      type: object
      required: [id, title, category, goal, owner, narratedSummary, urgency, deadline, expectedValue, attention, cost, risk, reversibility, confidence, recommendation, priorityReason, options, evidence, assumptions, unresolvedQuestions, simulations, participants, status]
      properties:
        id: { type: string }
        title: { type: string }
        category: { type: string }
        goal: { type: string }
        owner: { type: string }
        narratedSummary: { type: string }
        urgency: { type: string }
        deadline: { type: string }
        expectedValue: { type: string }
        attention: { type: string }
        cost: { type: string }
        risk: { type: string }
        reversibility: { type: string }
        confidence: { type: number }
        recommendation: { type: string }
        priorityReason: { type: string }
        options:
          type: array
          items: { $ref: "#/components/schemas/CruxOption" }
        evidence:
          type: array
          items: { $ref: "#/components/schemas/Evidence" }
        assumptions:
          type: array
          items: { type: string }
        unresolvedQuestions:
          type: array
          items: { type: string }
        simulations:
          type: array
          items: { $ref: "#/components/schemas/Simulation" }
        participants:
          type: array
          items: { $ref: "#/components/schemas/AgentParticipant" }
        status: { type: string, enum: [open, resolved, deploying, deployed] }
        selectedResolution: { type: ["string", "null"] }
        deploymentStatus: { type: ["string", "null"] }
    ResolutionInput:
      type: object
      required: [optionId]
      properties:
        optionId: { type: string }
    ConversationInput:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 1 }
        participant: { type: ["string", "null"] }
    ConversationReply:
      type: object
      required: [agent, message, confidenceShift]
      properties:
        agent: { type: string }
        message: { type: string }
        confidenceShift: { type: number }
...