# lib/api-zod/src/generated/types/agentParticipant.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AgentParticipant {
  name: string;
  role: string;
  initials: string;
  color: string;
}