# lib/api-zod/src/generated/types/crux.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentParticipant } from './agentParticipant';
import type { CruxOption } from './cruxOption';
import type { CruxStatus } from './cruxStatus';
import type { Evidence } from './evidence';
import type { Simulation } from './simulation';

export interface Crux {
  id: string;
  title: string;
  category: string;
  goal: string;
  owner: string;
  narratedSummary: string;
  urgency: string;
  deadline: string;
  expectedValue: string;
  attention: string;
  cost: string;
  risk: string;
  reversibility: string;
  confidence: number;
  recommendation: string;
  priorityReason: string;
  options: CruxOption[];
  evidence: Evidence[];
  assumptions: string[];
  unresolvedQuestions: string[];
  simulations: Simulation[];
  participants: AgentParticipant[];
  status: CruxStatus;
  /** @nullable */
  selectedResolution?: string | null;
  /** @nullable */
  deploymentStatus?: string | null;
}