# lib/api-zod/src/generated/types/evidence.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EvidenceSignal } from './evidenceSignal';

export interface Evidence {
  source: string;
  detail: string;
  signal: EvidenceSignal;
}