# lib/api-zod/src/generated/types/simulation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Simulation {
  label: string;
  probability: number;
  revenue: string;
  supportLoad: string;
  accent: string;
}